odoo_rfid_mqtt_client/requirements.txt
paho-mqtt
pycryptodome
